Add 23.11 Tapir artwork
This commit adds artwork to the NixOS 23.11 release featuring a Tapir.

The logo was taken from NixOS/nixos-artwork.

  https://github.com/NixOS/nixos-artwork/blob/master/logo/nix-snowflake.svg

The image of the Tapir was taken from Pexels, which allows the free use
of their pictures without credit, which is beneficial to the nature of
this kind of artwork.

  https://www.pexels.com/photo/malayan-tapir-15401340/

I have cropped the picture and turned down the saturation, because there
were discolorations at the nose and ears of the Tapir.