Add KDE Matrix icon + replace template room png with svg
The KDE Matrix room icon was added as a layer to the room template.
The colors used were selected to be unique and as close to the KDE
project as possible.

This commit also replaces the default Matrix room globe png with it's
correct svg image taken from their SDK.